initial rework of the integration / page / e2e tests
